Generalized_Linear_Models/requirements.txt
streamlit
pillow
numpy
pandas
matplotlib
statsmodels
scipy
seaborn